third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/overridemimetype-invalid-mime-type-expected.txt

This is a testharness.js-based test.
FAIL Bogus MIME type does not override encoding assert_equals: expected "text/html;charset=windows-1252" but got "bogus"
FAIL Bogus MIME type does not override encoding, 2 assert_equals: expected "ÿ" but got "\ufffd"
FAIL Bogus MIME type does override MIME type assert_equals: expected "text/xml" but got "bogus"
Harness: the test ran to completion.
